requirements.txt
streamlit>=1.31.0
pandas>=2.1.0
yfinance>=0.2.36
anthropic>=0.18.0
plotly>=5.18.0
scikit-learn>=1.4.0
python-dotenv>=1.0.1
numpy>=1.26.0
xgboost>=2.0.0
xgboost>=2.0.0
